+++ src/main/php/CaughtThrowable.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+/**
+ * This file is part of bovigo\assert.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace bovigo\assert;
+
+/**
+ * Allows to make assertions on a caught throwable.
+ *
+ * @since  8.1.0
+ */
+class CaughtThrowable extends CatchedException
+{
+    // intentionally empty
+}